Revision all id int to long

+++ src/main/java/com/springboot/moa/result/ResultDao.java
@@ -16,27 +16,27 @@ public class ResultDao {
     public void setDataSource(DataSource dataSource) {this.jdbcTemplate = new JdbcTemplate(dataSource);}
 
     // 앞 인자는 다른 곳에서 받아올 수 있는 PK (다른 Res 에서 반환 된 값)
-    public int insertResults(int postId, PostResultReq postResultReq) {
+    public long insertResults(long postId, PostResultReq postResultReq) {
         String insertResultQuery = "INSERT INTO result (post_id, user_id) VALUES (?, ?)";
         Object[] insertResultsParams = new Object[]{postId, postResultReq.getUserId()};
         this.jdbcTemplate.update(insertResultQuery, insertResultsParams);
         String lastInsertIdxQuery = "select last_insert_id()";
-        return this.jdbcTemplate.queryForObject(lastInsertIdxQuery, int.class);
+        return this.jdbcTemplate.queryForObject(lastInsertIdxQuery, long.class);
     }
 
-    public int insertResultDetails(int resultId, PostDetailResultReq postDetailResultReq) {
+    public long insertResultDetails(long resultId, PostDetailResultReq postDetailResultReq) {
         String insertResultDetailsQuery = "INSERT INTO result_detail (result_id, result) VALUES (?, ?)";
         Object[] insertResultDetailsParams = new Object[]{resultId, postDetailResultReq.getResult()};
         this.jdbcTemplate.update(insertResultDetailsQuery, insertResultDetailsParams);
         String lastInsertIdxQuery = "select last_insert_id()";
-        return this.jdbcTemplate.queryForObject(lastInsertIdxQuery, int.class);
+        return this.jdbcTemplate.queryForObject(lastInsertIdxQuery, long.class);
     }
 
 
-    public int selectPostPoint(int postId){
+    public int selectPostPoint(long postId){
         String selectPostPointQuery = "select point from post where post_id = ?";
-        int selectPostPointParam = postId;
+        long selectPostPointParam = postId;
 
-        return this.jdbcTemplate.queryForObject(selectPostPointQuery,int.class,selectPostPointParam);
+        return this.jdbcTemplate.queryForObject(selectPostPointQuery, int.class, selectPostPointParam);
     }
 }
